pyproject.toml
[project]
name = "forefree-vpn"
version = "0.1.0"
description = "Simple open-source VPN for learning networking fundamentals"
requires-python = ">=3.10"
dependencies = [
    "aws-cdk-lib>=2.220.0",
    "boto3>=1.40.58",
    "constructs>=10.4.2",
    "cryptography>=46.0.3",
    "python-dotenv",
]

[project.scripts]
pyvpn-client = "client.main:main"
pyvpn-server = "server.main:main"
setup-config = "common.setup_config:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["client", "server", "common"]
